+++ tests/MoneyTest.php
@@ -9,6 +9,7 @@
 
 use Cline\Math\BigDecimal;
 use Cline\Math\BigInteger;
+use Cline\Math\BigNumber;
 use Cline\Math\BigRational;
 use Cline\Math\Exception\DivisionByZeroException;
 use Cline\Math\Exception\NumberFormatException;
@@ -242,6 +243,10 @@ dataset('providerDividedBy', fn (): array => [
 ]);
 test('quotient and remainder', function (array $money, int $divisor, string $expectedQuotient, string $expectedRemainder): void {
     $money = Money::of(...$money);
+
+    self::assertMoneyIs($expectedQuotient, $money->quotient($divisor));
+    self::assertMoneyIs($expectedRemainder, $money->remainder($divisor));
+
     [$quotient, $remainder] = $money->quotientAndRemainder($divisor);
 
     self::assertMoneyIs($expectedQuotient, $quotient);
@@ -276,6 +281,10 @@ dataset('providerAllocate', fn (): array => [
     [['0.02', 'EUR'], [1, 1, 3, 1], ['EUR 0.01', 'EUR 0.00', 'EUR 0.01', 'EUR 0.00']],
     [[-100, 'USD'], [30, 20, 40, 40], ['USD -23.08', 'USD -15.39', 'USD -30.77', 'USD -30.76']],
     [['0.03', 'GBP'], [75, 25], ['GBP 0.03', 'GBP 0.00']],
+    [[100, 'USD'], ['30', BigNumber::of(20), '40'], ['USD 33.34', 'USD 22.22', 'USD 44.44']],
+    [[100, 'USD'], ['0.5', BigRational::of('3/2')], ['USD 25.00', 'USD 75.00']],
+    [[100, 'USD'], [BigRational::of('1/3'), BigRational::of('2/3')], ['USD 33.34', 'USD 66.66']],
+    [[100, 'USD'], [BigRational::of('1/3'), BigRational::of('2/7')], ['USD 53.85', 'USD 46.15']],
 ]);
 test('allocate empty list', function (): void {
     $money = Money::of(50, 'USD');
@@ -318,6 +327,10 @@ dataset('providerAllocateWithRemainder', fn (): array => [
     [['0.02', 'EUR'], [1, 1, 3, 1], ['EUR 0.00', 'EUR 0.00', 'EUR 0.00', 'EUR 0.00', 'EUR 0.02']],
     [[-100, 'USD'], [30, 20, 40, 40], ['USD -23.07', 'USD -15.38', 'USD -30.76', 'USD -30.76', 'USD -0.03']],
     [['0.03', 'GBP'], [75, 25], ['GBP 0.00', 'GBP 0.00', 'GBP 0.03']],
+    [[100, 'USD'], ['30', BigNumber::of(20), '40'], ['USD 33.33', 'USD 22.22', 'USD 44.44', 'USD 0.01']],
+    [['100.01', 'USD'], ['0.5', BigRational::of('3/2')], ['USD 25.00', 'USD 75.00', 'USD 0.01']],
+    [[100, 'USD'], [BigRational::of('1/3'), BigRational::of('2/3')], ['USD 33.33', 'USD 66.66', 'USD 0.01']],
+    [[100, 'USD'], [BigRational::of('1/3'), BigRational::of('2/7')], ['USD 53.83', 'USD 46.14', 'USD 0.03']],
 ]);
 test('allocate with remainder empty list', function (): void {
     $money = Money::of(50, 'USD');